+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## master (unreleased)
 
+### Changes
+
+* Update ChromeDriver to `87.0.4280.20`
+* Change test for `WebDriver#element_size_by_js`
+
 ## 0.6.3 (2020-09-11)
 
 ### New Features